refactor(contribute): decompose god-component into focused sub-components [CS-07]
Phase 10 (wave B): split contribute-client.tsx (313 -> 20 LOC composition root):
- use-contributors.ts: extract Contributor type + useContributors hook (fetch +
  EXCLUDED_USERS filter, graceful empty-on-failure). Called once in the root.
- contribute-hero.tsx: header + CTAs + founder quote (takes contributorCount).
- top-contributors.tsx: avatar cluster + name list.
- getting-started-steps.tsx: the 4 unique steps (data-driven STEPS table).
- sponsor-card.tsx: sponsor iframe block.

Verify: 408 tests pass, tsc clean, build green (/contribute route), lint 19->18.
